cobot320 code in python,
the code meant to run on the robot itself. 
seems to have problems in the forward kinematics as, calculations of forward kinematics on unity and python yeld different results with the same calculation steps.
otherwise it conyains diferent base formulas and functions aswell as an example of the inverse kinematic code.